Bayesian inference: completed refactor of python code
Since the hydro code api was modified to take command line
arguments, in an effort to promote parallel execution, the
python interface had to modified to exploit this.

Upon making these changes, I decided it was time to separate
out some of the logic I had stuck into the Bayesian analysis class
since this modularity would lend itself to nicer costumization

Work still needs to be done to formalize all the function comments
and a code an example, minimally functional product as a use case
for my code.