Intermittent commit: Still trying to debug sampler
The number of bins has as a substantial effect on the shape of
the distribution. This is unexpected behavior, which implies
a bug in the either plotting code or the sampler itself.

As a starting point, I am willing to assuem that the sampler
preforms correctly, but that instead the seaborn library plotting
rountines deform the underlying distribution, as is warned in the
documentation