Bayesian Inference: Major rewrite of mult-MCMC run
The idea of the module was to evaluate if the MCMC behaved as expected
However, the seaborn plotting library seem to distort the histograms
when plotting.
I removed the seaborn plotting routines form module, however, they
remain in the Bayesian inference class, as they are the standard method
for making corner plots.
Those interested can do a systematic study of the seaborn library vs
a self-written plotting rountine and determine what `smoothing
parameter` are needed to make everything behave nicely