Hydro Code: Adding addaptive time step to mvah and updated inverter
The numerical inverter takes many many steps to converge on large
values for xi. One way to speed up is to reduce dimensionality of
search to find Lambda and xi using PT/E and PL/E, as the ratios
cancel the fugacity dependence, and then 1-d line search for alpha.
Another way is to give better starting conditions for the parameters.
The ladder is the approach we implement.

With large xi, we encountered convergence issues with RK4 numerical
integrator. The time steps were to large and needed to be shrunk.
However, this comes at a severe performance cost, especially for
the MCMC. The solution is to incorporate a adaptive time step that
recursively divideds a single time step in 4 - 10, and performs
the time integration for these sub-intervals.
The ladder is the approach I have started to take, however there is
still work to be done.