Hydro Code: Finished adaptive RK4 method
The adaptive RK4 method does fairly well, it is not 100%
faithful to the exact solution evolution being off in the second digit
But for now I am not sure how to alleviate these concerns.

The viscous hydros preform twice as good as MVAH and VAH is off in the
fourth digit. So this problem needs to be addressed more rigorously.